--- src/data.js
@@ -1,18 +1,0 @@
-export const task = [
-    {
-        id: 1,
-        charge:"Sacar a caminar al perro"
-    },
-    {
-        id: 2,
-        charge:"Lavar la ropa"
-    },
-    {
-        id: 3,
-        charge:"Doblar la ropa"
-    },
-    {
-        id: 4,
-        charge:"Crear lista de tareas"
-    },
-  ];
